+++ libopendmi/include/opendmi/entity/intel/rsd-pcie.h
@@ -10,6 +10,7 @@
 #pragma once
 
 #include <opendmi/entity.h>
+#include <opendmi/entity/common.h>
 
 typedef struct dmi_intel_rsd_pcie dmi_intel_rsd_pcie_t;
 
@@ -22,7 +23,7 @@ struct dmi_intel_rsd_pcie
      * @brief PCI class code is used to identify the generic function of the
      * device.
      */
-    uint8_t pci_class;
+    dmi_pci_class_t pci_class;
 
     /**
      * @brief The physical slot number of the slot connected to the PCIe port.
@@ -30,27 +31,27 @@ struct dmi_intel_rsd_pcie
      * Slots (type 9) attribute structure for the slot this PCIe device is
      * located in.
      */
-    uint16_t pci_slot_number;
+    dmi_pci_slot_t pci_slot_number;
 
     /**
      * @brief Vendor identification number.
      */
-    uint16_t vendor_id;
+    dmi_pci_vendor_id_t vendor_id;
 
     /**
      * @brief Device identification number.
      */
-    uint16_t device_id;
+    dmi_pci_device_id_t device_id;
 
     /**
      * @brief Sub-vendor identifier used to differentiate specific model.
      */
-    uint16_t sub_vendor_id;
+    dmi_pci_vendor_id_t sub_vendor_id;
 
     /**
      * @brief Sub-device identifier used to differentiate specific model.
      */
-    uint16_t sub_device_id;
+    dmi_pci_device_id_t sub_device_id;
 
     /**
      * @brief Current link speed.